+++ src/main/java/de/tum/in/lrr/hmm/ISequence.java
@@ -6,7 +6,7 @@ import java.util.List;
  * @author <a href="bauerb@in.tum.de">Bernhard Bauer</a>
  *
  */
-public interface ISequence extends List<Integer> {
+public interface ISequence extends List<Byte> {
     int length();
 
     String getIdentifier();
