+++ src/main/java/de/tum/in/lrr/hmm/util/ByteArray.java
@@ -3,18 +3,18 @@ package de.tum.in.lrr.hmm.util;
 import java.util.Iterator;
 
 
-public class IntArray {
+public class ByteArray {
     public static final int DEFAULT_LIST_LENGTH = 16;
 
-    public static int[] forList(Iterator<Integer> it) {
+    public static byte[] forList(Iterator<Byte> it) {
         return forList(it, DEFAULT_LIST_LENGTH);
     }
 
-    public static int[] forList(Iterator<Integer> it, int length) {
-        int[] array = new int[length];
+    public static byte[] forList(Iterator<Byte> it, int length) {
+        byte[] array = new byte[length];
         int numItems = 0;
         while (it.hasNext()) {
-            final int i = it.next();
+            final byte i = it.next();
             if (numItems >= array.length) {
                 final int newSize;
                 if (array.length >= Integer.MAX_VALUE >> 1) {
@@ -22,7 +22,7 @@ public class IntArray {
                 } else {
                     newSize = array.length << 1;
                 }
-                final int[] newArray = new int[newSize];
+                final byte[] newArray = new byte[newSize];
                 System.arraycopy(array, 0, newArray, 0, array.length);
                 array = newArray;
             }
@@ -30,7 +30,7 @@ public class IntArray {
             numItems++;
         }
         if (array.length != numItems) {
-            final int[] newArray = new int[numItems];
+            final byte[] newArray = new byte[numItems];
             System.arraycopy(array, 0, newArray, 0, numItems);
             array = newArray;
         }
